setup a suite of detonations with Strang + SDC and NSE (#1169)
this cleans up the submission scripts, adds some docs, and slightly tweaks the initial conditions
