+++ Exec/science/Detonation/GNUmakefile.nse
@@ -1,0 +1,31 @@
+PRECISION  = DOUBLE
+PROFILE    = FALSE
+
+DEBUG      = FALSE
+
+DIM        = 1
+
+COMP	   = gnu
+
+USE_MPI    = TRUE
+USE_OMP    = FALSE
+
+USE_REACT  = TRUE
+
+USE_PROB_PARAMS = TRUE
+
+CASTRO_HOME = ../../..
+
+GPU_COMPATIBLE_PROBLEM = TRUE
+
+# This sets the EOS directory in $(MICROPHYSICS_HOME)/eos
+EOS_DIR     := helmholtz
+
+# This sets the EOS directory in $(MICROPHYSICS_HOME)/networks
+NETWORK_DIR := aprox19
+USE_NSE = TRUE
+
+Bpack   := ./Make.package
+Blocs   := .
+
+include $(CASTRO_HOME)/Exec/Make.Castro
