+++ Exec/science/Detonation/inputs-det-x.nse
@@ -1,13 +1,14 @@
 # ------------------  INPUTS TO MAIN PROGRAM  -------------------
 max_step = 100000
-stop_time =  3.e-3
+stop_time =  12.e-5
 
 # PROBLEM SIZE & GEOMETRY
 geometry.is_periodic = 0 0 0
 geometry.coord_sys   = 0  # 0 => cart, 1 => RZ  2=>spherical
-geometry.prob_lo     = 0        0    0
-geometry.prob_hi     = 6.5536e7
-amr.n_cell           = @@NZONES@@
+geometry.prob_lo     = 0     0     0
+geometry.prob_hi     = 6.5536e5
+amr.n_cell           = 256
+
 
 # >>>>>>>>>>>>>  BC FLAGS <<<<<<<<<<<<<<<<
 # 0 = Interior           3 = Symmetry
@@ -32,14 +33,17 @@ castro.small_temp     = 1.e7
 
 castro.riemann_solver = 0
 
-castro.time_integration_method = @@method@@
-castro.sdc_iters = @@SDC_ITERS@@
+castro.time_integration_method = 0
 
 # TIME STEP CONTROL
-castro.cfl            = @@CFL@@     # cfl number for hyperbolic system
+castro.cfl            = 0.5     # cfl number for hyperbolic system
 castro.init_shrink    = 0.1     # scale back initial timestep
 castro.change_max     = 1.05    # scale back initial timestep
 
+
+castro.dtnuc_e = 0.25
+castro.use_retry = 1
+castro.max_subcycles = 32
 
 # DIAGNOSTICS & VERBOSITY
 castro.sum_interval   = 1       # timesteps between computing mass
@@ -47,7 +51,7 @@ castro.v              = 1       # verbosity in Castro.cpp
 amr.v                 = 1       # verbosity in Amr.cpp
 #amr.grid_log        = grdlog  # name of grid logging file
 
-# REFINEMENT / REGRIDDING
+# REFINEMENT / REGRIDDING 
 amr.max_level       = 0       # maximum level number allowed
 amr.ref_ratio       = 2 2 2 2 # refinement ratio
 amr.regrid_int      = 2 2 2 2 # how often to regrid
@@ -61,8 +65,8 @@ amr.check_int       = 1000         # number of timesteps between checkpoints
 
 # PLOTFILES
 amr.plot_file       = det_x_plt  # root name of plotfile
-amr.plot_per = 5.e-4
+amr.plot_per = 5.e-6
 amr.derive_plot_vars = ALL
 
 #PROBIN FILENAME
-amr.probin_file = probin.nse_test.sdc
+amr.probin_file = probin-det-x.nse_disabled
