--- Exec/science/Detonation/nse_runs/probin.nse_test.strang
@@ -1,45 +1,0 @@
-&fortin
-  T_l = 4.d9
-  T_r = 5.d7
-
-  dens = 2.d8
-  cfrac = 0.d0
-  
-  smallx = 1.e-10
-
-  idir = 1
-
-  w_T = 5.d-4
-  center_T = 0.1
-/
-
-&tagging
-  denerr = 3
-  dengrad = 0.01
-  max_denerr_lev = 0
-  max_dengrad_lev = 0
-
-  presserr = 3
-  pressgrad = 0.01
-  max_presserr_lev = 0
-  max_pressgrad_lev = 0
-
-  temperr = 4.e9
-  tempgrad = 1.e8
-  max_temperr_lev = 5
-  max_tempgrad_lev = 5
-/
-
-&extern
-   call_eos_in_rhs = T
-   do_constant_volume_burn = T
-
-   rtol_spec = 1.e-10
-   rtol_enuc = 1.e-10
-   rtol_temp = 1.e-10
-   use_tables = F
-   jacobian = 2
-   ode_max_steps = 1500000
-
-   abort_on_failure = F
-/
