+++ Exec/science/Detonation/nse_runs/setup_runs.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import os
 import shutil
 import subprocess
@@ -5,16 +7,17 @@ import shlex
 from multiprocessing import Pool
 import time
 
-CFL = [0.8, 0.4, 0.2, 0.1]
+CFL = [0.5, 0.25, 0.1]
 SDC_ITERS = [2, 3]
-DTNUC_E = [1.e200, 0.5]
+DTNUC_E = [1.e200, 0.25, 0.1]
 
-NZONES = [256, 512, 1024, 2048, 4096, 8192]
+NZONES = [256, 512, 1024, 2048] #, 8192]
 
 job_list = []
 
-NEEDED_SDC_FILES = ["helm_table.dat", "probin.nse_test.sdc", "Castro1d.gnu.SDC.ex"]
-NEEDED_STRANG_FILES = ["helm_table.dat", "probin.nse_test.strang", "Castro1d.gnu.ex"]
+COMMON_FILES = ["helm_table.dat", "probin-det-x.nse_disabled", "nse19.tbl"]
+NEEDED_SDC_FILES = ["Castro1d.gnu.MPI.SMPLSDC.ex"] + COMMON_FILES
+NEEDED_STRANG_FILES = ["Castro1d.gnu.MPI.ex"] + COMMON_FILES
 
 def setup_runs():
 
@@ -58,30 +61,36 @@ def setup_runs():
 
     # setup the simplified SDC runs
     for nz in NZONES:
-        for c in CFL:
-            for niter in SDC_ITERS:
+        for dtn in DTNUC_E:
+            for c in CFL:
+                for niter in SDC_ITERS:
 
-                # make the output directory
-                odir = "det_sdc_iter{}_cfl{}_nzones{}".format(niter, c, nz)
-                os.mkdir(odir)
+                    # don't do all the CFLs when using the nuc energy dt limiter
+                    if dtn < 1.0 and c != CFL[0]:
+                        continue
 
-                job_list.append(odir)
+                    # make the output directory
+                    odir = "det_sdc_iter{}_cfl{}_dtnuce{}_nzones{}".format(niter, c, dtn, nz)
+                    os.mkdir(odir)
 
-                # dump the metdata file
-                with open("{}/run.meta".format(odir), "w") as meta:
-                    meta.write("cfl = {}\n".format(c))
-                    meta.write("nzones = {}\n".format(nz))
-                    meta.write("integrator = SDC\n")
-                    meta.write("niters = {}\n".format(niter))
+                    job_list.append(odir)
 
-                # write the inputs file
-                with open("{}/inputs".format(odir), "w") as inpf:
-                    for line in sdc_template:
-                        inpf.write(line.replace("@@CFL@@", str(c)).replace("@@NZONES@@", str(nz)).replace("@@SDC_ITERS@@", str(niter)).replace("@@method@@", "3"))
+                    # dump the metdata file
+                    with open("{}/run.meta".format(odir), "w") as meta:
+                        meta.write("cfl = {}\n".format(c))
+                        meta.write("nzones = {}\n".format(nz))
+                        meta.write("integrator = SDC\n")
+                        meta.write("niters = {}\n".format(niter))
+                        meta.write("dtnuc_e = {}\n".format(dtn))
 
-                # copy the remaining files
-                for f in NEEDED_SDC_FILES:
-                    shutil.copy(f, odir)
+                    # write the inputs file
+                    with open("{}/inputs".format(odir), "w") as inpf:
+                        for line in sdc_template:
+                            inpf.write(line.replace("@@CFL@@", str(c)).replace("@@NZONES@@", str(nz)).replace("@@SDC_ITERS@@", str(niter)).replace("@@method@@", "3").replace("@@DTNUC_E@@", str(dtn)))
+
+                    # copy the remaining files
+                    for f in NEEDED_SDC_FILES:
+                        shutil.copy(f, odir)
 
     return job_list
 
@@ -103,9 +112,9 @@ def do_run(name):
     print("running {}...".format(name))
 
     if "sdc" in name:
-        command = "./Castro1d.gnu.SDC.ex inputs"
+        command = "./Castro1d.gnu.MPI.SMPLSDC.ex inputs"
     else:
-        command = "./Castro1d.gnu.ex inputs"
+        command = "./Castro1d.gnu.MPI.ex inputs"
 
     cwd = os.getcwd()
     os.chdir(name)
@@ -128,5 +137,5 @@ if __name__ == "__main__":
     job_list = setup_runs()
 
     # do the runs
-    p = Pool(16)
+    p = Pool(8)
     p.map(do_run, job_list)
